Add files via upload
This project uses AutoML (AutoGluon) to predict survival outcomes in the Titanic dataset. It trains a model on the training data, evaluates performance using a precision-recall curve (achieving an AP of 0.80), and selects an optimal threshold (0.415) based on the F1 score. The final model achieves a precision of 0.769, recall of 0.725, and F1 score of 0.746 on the validation set. Predictions are generated for the test set and saved for Kaggle submission. Includes data preprocessing, model training, evaluation, and visualization in a Jupyter Notebook.